# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample fiber project server.",
        "title": "Machine task",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Authenticate a user and return a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "userLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
...
        "/signup": {
            "post": {
                "description": "Signup a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a new user account",
                "parameters": [
                    {
                        "description": "User signup data",
                        "name": "userProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UserSignup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "account created",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requestmodels.UserLogin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 5
                }
            }
        },
        "requestmodels.UserSignup": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 5
                },
                "profileHeadline": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
        "responsemodels.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "result": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "authorization": {
            "description": "IMPORTANT: TYPE \"Bearer\" FOLLOWED BY A SPACE AND JWT TOKEN.",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}